Add note and poll template forms to event creator
